[Fix/#35] DetailView 및 커뮤니티 상세화면 수정
