+++ GLINT-iOS/GLINT-iOS/Source/Presentation/Feature/TabBar/Components/CommunityTab.swift
@@ -24,6 +24,8 @@ struct CommunityTab: View {
         switch route {
         case .community:
             CommunityView()
+        case .communityDetail(let id):
+            CommunityDetailView(id: id)
         }
     }
 }
